Update: 댓글 삭제 시 사용자 인증 추가
사용자 인증 로직 구현 및 front에서 댓글 작성자와 유저가 같을 경우만 삭제 버튼 출력

+++ DU/src/main/java/du/reply/controller/ReplyController.java
@@ -1,5 +1,9 @@
 package du.reply.controller;
 
+import java.io.IOException;
+import java.io.PrintWriter;
+
+import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,6 +12,7 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import du.board.domain.BoardVO;
 import du.reply.domain.ReplyVO;
 import du.reply.service.ReplyService;
 import du.user.domain.UserVO;
@@ -34,10 +39,34 @@ public class ReplyController {
 	}
 
 	@RequestMapping("/replyDelete.do")
-	public String replyDelete(HttpSession session, @RequestParam long idx, @RequestParam String boardIdx) {
-		
-		replyService.deleteReply(idx);
-		
-		return "redirect:/boardInfoPage/" + boardIdx + ".do";
+	public String replyDelete(HttpSession session, HttpServletResponse response, @RequestParam long idx, @RequestParam String boardIdx) {
+
+		ReplyVO reply = replyService.selectReply(idx);
+		UserVO user = (UserVO) session.getAttribute("USER");
+
+		if (user == null) {
+			return "redirect:/loginPage.do";
+		} else if (reply.getWriterId().equals(user.getUserId())) {
+
+			replyService.deleteReply(idx);
+
+			return "redirect:/boardInfoPage/" + boardIdx + ".do";
+
+		} else {
+			response.setContentType("text/html; charset=UTF-8");
+
+			try {
+
+				PrintWriter out = response.getWriter();
+				out.println("<script>alert('권한이 없습니다.'); history.back();</script>");
+				out.flush();
+
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
+
+			return null;
+		}
+
 	}
 }
